refactor: lineManagerView.js삭제 후 lineView.js에서 하도록 변경
